extract native-stream-sink and source into their own files
