+++ c-dependencies/js-compute-runtime/js-compute-builtins.cpp
@@ -33,6 +33,8 @@
 #include "builtin.h"
 #include "builtins/env.h"
 #include "builtins/logger.h"
+#include "builtins/native-stream-sink.h"
+#include "builtins/native-stream-source.h"
 
 using JS::CallArgs;
 using JS::CallArgsFromVp;
@@ -702,23 +704,6 @@ static bool create(JSContext *cx, HandleObject global) {
   return JS_DefineFunctions(cx, crypto, methods);
 }
 } // namespace Crypto
-
-namespace NativeStreamSource {
-typedef bool PullAlgorithm(JSContext *cx, CallArgs args, HandleObject stream, HandleObject owner,
-                           HandleObject controller);
-typedef bool CancelAlgorithm(JSContext *cx, CallArgs args, HandleObject stream, HandleObject owner,
-                             HandleValue reason);
-
-JSObject *create(JSContext *cx, HandleObject owner, HandleValue startPromise, PullAlgorithm *pull,
-                 CancelAlgorithm *cancel);
-static JSObject *get_stream_source(JSContext *cx, HandleObject stream);
-JSObject *owner(JSObject *self);
-JSObject *stream(JSObject *self);
-bool stream_has_native_source(JSContext *cx, HandleObject stream);
-bool stream_is_body(JSContext *cx, HandleObject stream);
-bool lock_stream(JSContext *cx, HandleObject stream);
-JSObject *piped_to_transform_stream(JSObject *source);
-} // namespace NativeStreamSource
 
 enum class BodyReadResult {
   ArrayBuffer,
@@ -1538,332 +1523,6 @@ bool body_get(JSContext *cx, CallArgs args, HandleObject self, bool create_if_un
   return true;
 }
 } // namespace RequestOrResponse
-
-namespace NativeStreamSink {
-static JSObject *get_stream_sink(JSContext *cx, HandleObject stream);
-JSObject *owner(JSObject *self);
-} // namespace NativeStreamSink
-
-// A JS class to use as the underlying source for native readable streams, used
-// for Request/Response bodies and TransformStream.
-namespace NativeStreamSource {
-namespace Slots {
-enum {
-  Owner,          // Request or Response object, or TransformStream.
-  Controller,     // The ReadableStreamDefaultController.
-  InternalReader, // Only used to lock the stream if it's consumed internally.
-  StartPromise,   // Used as the return value of `start`, can be undefined.
-                  // Needed to properly implement TransformStream.
-  PullAlgorithm,
-  CancelAlgorithm,
-  PipedToTransformStream, // The TransformStream this source's stream is piped
-                          // to, if any. Only applies if the source backs a
-                          // RequestOrResponse's body.
-  Count
-};
-};
-
-bool is_instance(JSObject *obj);
-
-JSObject *owner(JSObject *self) {
-  MOZ_ASSERT(is_instance(self));
-  return &JS::GetReservedSlot(self, Slots::Owner).toObject();
-}
-
-JSObject *stream(JSObject *self) { return RequestOrResponse::body_stream(owner(self)); }
-
-Value startPromise(JSObject *self) {
-  MOZ_ASSERT(is_instance(self));
-  return JS::GetReservedSlot(self, Slots::StartPromise);
-}
-
-PullAlgorithm *pullAlgorithm(JSObject *self) {
-  MOZ_ASSERT(is_instance(self));
-  return (PullAlgorithm *)JS::GetReservedSlot(self, Slots::PullAlgorithm).toPrivate();
-}
-
-CancelAlgorithm *cancelAlgorithm(JSObject *self) {
-  MOZ_ASSERT(is_instance(self));
-  return (CancelAlgorithm *)JS::GetReservedSlot(self, Slots::CancelAlgorithm).toPrivate();
-}
-
-JSObject *controller(JSObject *self) {
-  MOZ_ASSERT(is_instance(self));
-  return &JS::GetReservedSlot(self, Slots::Controller).toObject();
-}
-
-/**
- * Returns the underlying source for the given controller iff it's an object,
- * nullptr otherwise.
- */
-static JSObject *get_controller_source(JSContext *cx, HandleObject controller) {
-  RootedValue source(cx);
-  bool success __attribute__((unused));
-  success = JS::ReadableStreamControllerGetUnderlyingSource(cx, controller, &source);
-  MOZ_ASSERT(success);
-  return source.isObject() ? &source.toObject() : nullptr;
-}
-
-static JSObject *get_stream_source(JSContext *cx, HandleObject stream) {
-  MOZ_ASSERT(JS::IsReadableStream(stream));
-  RootedObject controller(cx, JS::ReadableStreamGetController(cx, stream));
-  return get_controller_source(cx, controller);
-}
-
-bool stream_has_native_source(JSContext *cx, HandleObject stream) {
-  JSObject *source = get_stream_source(cx, stream);
-  return is_instance(source);
-}
-
-bool stream_is_body(JSContext *cx, HandleObject stream) {
-  JSObject *stream_source = get_stream_source(cx, stream);
-  return NativeStreamSource::is_instance(stream_source) &&
-         RequestOrResponse::is_instance(owner(stream_source));
-}
-
-void set_stream_piped_to_ts_writable(JSContext *cx, HandleObject stream, HandleObject writable) {
-  RootedObject source(cx, NativeStreamSource::get_stream_source(cx, stream));
-  MOZ_ASSERT(is_instance(source));
-  RootedObject sink(cx, NativeStreamSink::get_stream_sink(cx, writable));
-  RootedObject transform_stream(cx, NativeStreamSink::owner(sink));
-  MOZ_ASSERT(transform_stream);
-  JS::SetReservedSlot(source, Slots::PipedToTransformStream, ObjectValue(*transform_stream));
-}
-
-JSObject *piped_to_transform_stream(JSObject *self) {
-  MOZ_ASSERT(is_instance(self));
-  return JS::GetReservedSlot(self, Slots::PipedToTransformStream).toObjectOrNull();
-}
-
-bool lock_stream(JSContext *cx, HandleObject stream) {
-  MOZ_ASSERT(JS::IsReadableStream(stream));
-
-  bool locked;
-  JS::ReadableStreamIsLocked(cx, stream, &locked);
-  if (locked) {
-    JS_ReportErrorLatin1(cx, "Can't lock an already locked ReadableStream");
-    return false;
-  }
-
-  RootedObject self(cx, get_stream_source(cx, stream));
-  MOZ_ASSERT(is_instance(self));
-
-  auto mode = JS::ReadableStreamReaderMode::Default;
-  RootedObject reader(cx, JS::ReadableStreamGetReader(cx, stream, mode));
-  if (!reader)
-    return false;
-
-  JS::SetReservedSlot(self, Slots::InternalReader, JS::ObjectValue(*reader));
-  return true;
-}
-
-const unsigned ctor_length = 0;
-
-bool check_receiver(JSContext *cx, HandleValue receiver, const char *method_name);
-
-bool start(JSContext *cx, unsigned argc, Value *vp) {
-  METHOD_HEADER(1)
-
-  MOZ_ASSERT(args[0].isObject());
-  RootedObject controller(cx, &args[0].toObject());
-  MOZ_ASSERT(get_controller_source(cx, controller) == self);
-
-  JS::SetReservedSlot(self, Slots::Controller, args[0]);
-
-  // For TransformStream, StartAlgorithm returns the same Promise for both the
-  // readable and writable stream. All other native initializations of
-  // ReadableStream have StartAlgorithm return undefined. Instead of introducing
-  // both the StartAlgorithm as a pointer and startPromise as a value, we just
-  // store the latter or undefined, and always return it.
-  args.rval().set(startPromise(self));
-  return true;
-}
-
-bool pull(JSContext *cx, unsigned argc, Value *vp) {
-  METHOD_HEADER(1)
-
-  RootedObject owner(cx, NativeStreamSource::owner(self));
-  RootedObject controller(cx, &args[0].toObject());
-  MOZ_ASSERT(controller == NativeStreamSource::controller(self));
-  MOZ_ASSERT(get_controller_source(cx, controller) == self.get());
-
-  PullAlgorithm *pull = pullAlgorithm(self);
-  return pull(cx, args, self, owner, controller);
-}
-
-bool cancel(JSContext *cx, unsigned argc, Value *vp) {
-  METHOD_HEADER(0)
-
-  RootedObject owner(cx, NativeStreamSource::owner(self));
-  HandleValue reason(args.get(0));
-
-  CancelAlgorithm *cancel = cancelAlgorithm(self);
-  return cancel(cx, args, self, owner, reason);
-}
-
-const JSFunctionSpec methods[] = {JS_FN("start", start, 1, 0), JS_FN("pull", pull, 1, 0),
-                                  JS_FN("cancel", cancel, 1, 0), JS_FS_END};
-
-const JSPropertySpec properties[] = {JS_PS_END};
-
-CLASS_BOILERPLATE_NO_CTOR(NativeStreamSource)
-
-JSObject *create(JSContext *cx, HandleObject owner, HandleValue startPromise, PullAlgorithm *pull,
-                 CancelAlgorithm *cancel) {
-  RootedObject source(cx, JS_NewObjectWithGivenProto(cx, &class_, proto_obj));
-  if (!source)
-    return nullptr;
-
-  JS::SetReservedSlot(source, Slots::Owner, JS::ObjectValue(*owner));
-  JS::SetReservedSlot(source, Slots::StartPromise, startPromise);
-  JS::SetReservedSlot(source, Slots::PullAlgorithm, JS::PrivateValue((void *)pull));
-  JS::SetReservedSlot(source, Slots::CancelAlgorithm, JS::PrivateValue((void *)cancel));
-  JS::SetReservedSlot(source, Slots::PipedToTransformStream, JS::NullValue());
-  return source;
-}
-} // namespace NativeStreamSource
-
-// A JS class to use as the underlying sink for native writable streams, used
-// for TransformStream.
-namespace NativeStreamSink {
-typedef bool WriteAlgorithm(JSContext *cx, CallArgs args, HandleObject stream, HandleObject owner,
-                            HandleValue chunk);
-typedef bool AbortAlgorithm(JSContext *cx, CallArgs args, HandleObject stream, HandleObject owner,
-                            HandleValue reason);
-typedef bool CloseAlgorithm(JSContext *cx, CallArgs args, HandleObject stream, HandleObject owner);
-namespace Slots {
-enum {
-  Owner,          // TransformStream.
-  Controller,     // The WritableStreamDefaultController.
-  InternalWriter, // Only used to lock the stream if it's consumed internally.
-  StartPromise,   // Used as the return value of `start`, can be undefined.
-                  // Needed to properly implement TransformStream.
-  WriteAlgorithm,
-  AbortAlgorithm,
-  CloseAlgorithm,
-  // AbortAlgorithm, TODO: implement
-  Count
-};
-};
-
-bool is_instance(JSObject *obj);
-
-JSObject *owner(JSObject *self) { return &JS::GetReservedSlot(self, Slots::Owner).toObject(); }
-
-Value startPromise(JSObject *self) { return JS::GetReservedSlot(self, Slots::StartPromise); }
-
-WriteAlgorithm *writeAlgorithm(JSObject *self) {
-  return (WriteAlgorithm *)JS::GetReservedSlot(self, Slots::WriteAlgorithm).toPrivate();
-}
-
-AbortAlgorithm *abortAlgorithm(JSObject *self) {
-  return (AbortAlgorithm *)JS::GetReservedSlot(self, Slots::AbortAlgorithm).toPrivate();
-}
-
-CloseAlgorithm *closeAlgorithm(JSObject *self) {
-  return (CloseAlgorithm *)JS::GetReservedSlot(self, Slots::CloseAlgorithm).toPrivate();
-}
-
-JSObject *controller(JSObject *self) {
-  return &JS::GetReservedSlot(self, Slots::Controller).toObject();
-}
-
-/**
- * Returns the underlying sink for the given controller iff it's an object,
- * nullptr otherwise.
- */
-static JSObject *get_controller_sink(JSContext *cx, HandleObject controller) {
-  RootedValue sink(cx, JS::WritableStreamControllerGetUnderlyingSink(cx, controller));
-  return sink.isObject() ? &sink.toObject() : nullptr;
-}
-
-static JSObject *get_stream_sink(JSContext *cx, HandleObject stream) {
-  RootedObject controller(cx, JS::WritableStreamGetController(cx, stream));
-  return get_controller_sink(cx, controller);
-}
-
-bool stream_has_native_sink(JSContext *cx, HandleObject stream) {
-  MOZ_RELEASE_ASSERT(JS::IsWritableStream(stream));
-
-  JSObject *sink = get_stream_sink(cx, stream);
-  return is_instance(sink);
-}
-
-const unsigned ctor_length = 0;
-
-bool check_receiver(JSContext *cx, HandleValue receiver, const char *method_name);
-
-bool start(JSContext *cx, unsigned argc, Value *vp) {
-  METHOD_HEADER(1)
-
-  MOZ_ASSERT(args[0].isObject());
-  RootedObject controller(cx, &args[0].toObject());
-  MOZ_ASSERT(get_controller_sink(cx, controller) == self);
-
-  JS::SetReservedSlot(self, Slots::Controller, args[0]);
-
-  // For TransformStream, StartAlgorithm returns the same Promise for both the
-  // readable and writable stream. All other native initializations of
-  // WritableStream have StartAlgorithm return undefined.
-  //
-  // Instead of introducing both the StartAlgorithm as a pointer and
-  // startPromise as a value, we just store the latter or undefined, and always
-  // return it.
-  args.rval().set(startPromise(self));
-  return true;
-}
-
-bool write(JSContext *cx, unsigned argc, Value *vp) {
-  METHOD_HEADER(1)
-
-  RootedObject owner(cx, NativeStreamSink::owner(self));
-  HandleValue chunk(args[0]);
-
-  WriteAlgorithm *write = writeAlgorithm(self);
-  return write(cx, args, self, owner, chunk);
-}
-
-bool abort(JSContext *cx, unsigned argc, Value *vp) {
-  METHOD_HEADER(1)
-
-  RootedObject owner(cx, NativeStreamSink::owner(self));
-  HandleValue reason(args[0]);
-
-  AbortAlgorithm *abort = abortAlgorithm(self);
-  return abort(cx, args, self, owner, reason);
-}
-
-bool close(JSContext *cx, unsigned argc, Value *vp) {
-  METHOD_HEADER(0)
-
-  RootedObject owner(cx, NativeStreamSink::owner(self));
-
-  CloseAlgorithm *close = closeAlgorithm(self);
-  return close(cx, args, self, owner);
-}
-
-const JSFunctionSpec methods[] = {JS_FN("start", start, 1, 0), JS_FN("write", write, 2, 0),
-                                  JS_FN("abort", abort, 2, 0), JS_FN("close", close, 1, 0),
-                                  JS_FS_END};
-
-const JSPropertySpec properties[] = {JS_PS_END};
-
-CLASS_BOILERPLATE_NO_CTOR(NativeStreamSink)
-
-JSObject *create(JSContext *cx, HandleObject owner, HandleValue startPromise, WriteAlgorithm *write,
-                 CloseAlgorithm *close, AbortAlgorithm *abort) {
-  RootedObject sink(cx, JS_NewObjectWithGivenProto(cx, &class_, proto_obj));
-  if (!sink)
-    return nullptr;
-
-  JS::SetReservedSlot(sink, Slots::Owner, JS::ObjectValue(*owner));
-  JS::SetReservedSlot(sink, Slots::StartPromise, startPromise);
-  JS::SetReservedSlot(sink, Slots::WriteAlgorithm, JS::PrivateValue((void *)write));
-  JS::SetReservedSlot(sink, Slots::AbortAlgorithm, JS::PrivateValue((void *)abort));
-  JS::SetReservedSlot(sink, Slots::CloseAlgorithm, JS::PrivateValue((void *)close));
-  return sink;
-}
-} // namespace NativeStreamSink
 
 namespace ReadableStream_additions {
 static PersistentRooted<JSObject *> proto_obj;
